reformat centernet
